requirements.txt

pandas>=2.0.0
numpy>=1.23.0
scikit-learn>=1.2.0
scipy>=1.10.0
matplotlib>=3.7.0
joblib>=1.2.0
pyyaml>=6.0
streamlit>=1.32.0
